+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.27.0
  - feat: adding `buttonWidth` attribute to `NesButton.text` constructor.
+ - feat: adding `iconSize` attribute to `NesButton.icon` constructor.
 
 # 0.26.0
  - feat: adding `NesIcons.lock`.